distribute prototyping tests to more files, activate more ignored filter tests
